Build Config Fixups
We ask for phonon in src.pro, but it is NOT required, this
is because the video playback was rewritten to use the vlc
libraries.

Additionally, the build.pro now asks for ordered builds to
ensure the qwt libs are built before src is started. This is
useful when performing builds with multipe jobs (make -j n).