PfPv plot adjust axis for track sprinters
Higher values for AEPF/CPV have been observed with
track sprinters. The fixed axis on the PfPv (QA)
charts truncate the display.

This patch adjusts the axes to suit the data and will
also ensure the zone shading and labels are extended
to the new maximums.

The PfPv code also contained a lot of tab characters
and the general style was quite dense, making it
difficult to read and understand. I have adjusted
to use a 4 space tab-stop and adjusted to match the
general style of the rest of GC.

Fixes #463.