Initial 'proper' support for Video Playback
The code to select media and then start/stop/pause
playback during a workout has been implemented.

This requires libvlc to be installed at compile time
and will also require Videolan (and desired codecs)
installed at runtime.

Since the main headaches are likely to be related to
deployment rather than coding this update has been
pushed for deployment support to be enhanced and
tested before closing the feature request on the
bug tracker.