Support Garmin USB2 ANT+ sticks on Mac OS X
This update completes support for USB2 sticks for ANT+ devices
in train view. The code now works on Linux and Mac as well as
Windows.

IOKit is now a required framework for builds on Mac OSX and
src.pro has been updated accordingly.

I have also added a tarball of the required version of libusb
in a new 'contrib' directory. Users may wish to use a later
version or download from an alternative source, but this
version is guaranteed to work.

Fixes #310