Fix MAC_OS_X_VERSION_MAX_ALLOWED
LionFullScreen used this macro to determine if the version
of the SDK is recent enough (Lion or higher).

Unfortunately, I forgot to also check if the OS was Mac OSX
and didn't include the correct header (AvailaibilityMacros.h)
to get the define. As a result all builds on anything other
than OSX Lion failed.