LionFullScreen saga. Ugh.
Even if the C pre-processor will remove a class definition
in a header file the QT moc utility will still create a
moc_Xxxx.cpp file which requires the class to be defined.

As a result the attempt to create an empty header for the
LionFullScreen class when compiling on earlier versions than
Lion causes issues.

The LionFullScreen class is now always declared and defined
but will be not do anything useful on earlier versions of
Mac OS X than Lion.