LibUsb does not have usb_reset on Windows
We call usb_reset between sessions on Linux/Mac
because the USB device is left in an undetermined
state after close. In the comments I noted that
it was odd this was not required on Windows.

Well, it turns out it is not available on Windows.
This patch ensures we do not try to call it when
compiling for Windows.