ToolBar even more like Flare
The "tabs" of the toolbar are now offset to the right
and fixed with a specific spacing.

Similarly, the training view buttons are fixed to the
left.

This ensures that the "tabs" don't move as toolbar
buttons are added/removed.

This removes the flickery artefacts when moving away
from train view and also gets ready for more specific
buttons on the other views.