Plot marker default color now Gray
Because black does not show up against a black background
(the default for a ride plot).

This was not a big issue, but it is used by the train view
as the color for the timer and results in black writing on
a black background.

This might be cool if you're in a band called Disaster Area
but not so cool if you want to ride your turbo.