Mac Video use GC_HAVE_LION
When compiling the Mac video stuff on Lion with SDK10.5
the code used a pre-processor directive to include code
based upon the max version allowed.

But if you are using an earlier SDK the max version allowed
is still higher, so instead we use the GC_HAVE_LION define
instead.

This change only really affects anyone on Lion or higher building
with an earlier version of the SDK.