Linux/Win disable screensaver using VLC
VLC will disable the screensaver during playback. This patch
exploits this functionality bringing it in line with the Mac
builds.

One small nit remains; under Windows/Linux the screensaver
is only disabled if video playback is being used, whilst on
the Mac the screensaver is disabled whenever you are active
(e.g. on a basic workout or streetview).