Fix Crash in ErgFilePlot TimeScaleDraw
The TimeScaleDraw declaration and definition in
ErgPlot.h clashes (silently) with the definition
in PerfPlot.h.

This caused a SEGV at runtime, but only with some
build tools (but notably Xcode on Lion). Not at
all sure how this has not arisen before.

ErgFilePlot now declares as HourTimeScaleDraw to
avoid the clash.