Fix FullScreen Crash on Mac
When building with the latest Xcode their is a conflict
between the QtMacSegmentedButton cocoa container and the
same declared in QtMacVideoWindow.

This patch removes QtMacSegmentedButton since we do not
use it.  There appears to be a container in 4.7 anyway
so we no longer declare our own.

Video and FullScreen are fully operational now on 10.6
or higher using 4.7.4 or higher of QT.