Introduce OperatingSystem global variable
The LibUsb (and possibly other classes) need to know
what operating system we are running on at runtime. In
the past this has been achieved with lots of #ifdef pre-
processor conditionals, but this makes the code quite
difficult to read in places.

This patch introduces the variable OperatingSystem and
it can be one of WINDOWS, LINUX, OSX.