Fix Histogram Crash
When selecting the speed tab on an imported ride-file, GC exits
with the following error on the command- line: ASSERT failure in
QVector<T>::operator[]: "index out of range", file
/usr/include/qt4/QtCore/qvector.h, line 347.

File: PowerHist.cpp on line 393
low has value -5.55111512312578e-17 which is like -0.0000000000x

This caused an out of bound error on array lookup.

Fixes #551.