Rolling Average Power/Cadence/HR tweaks
The slider in the controls should only be shown
for HR/Power/Cadence. Additionally, it is more
of a 'smoothing' (rolling) average, so the text
has been changed to reflect that.

This is to avoid confusion with the other series
entitled Average Power. Average Cadence and
Average heart rate - which are averages for the
entire ride.