Training Dials have own Color setting
For expoediency the training dials used the qwtplotmarker
color but this often lead to things like lap, time being
displayed in black on a black background.

We now have a realtime dials default color setting in the
appearances pane. The default is gray, so will show up on
a black or white background.