Set default workout directory for Train view
Since it can be confusing to new or occasional users, the
workout directory now defaults to the top-level GC directory.

This kind of makes sense since it is shared across all cyclists
and is guaranteed to exist, and will still work when running
from a USB stick.

When we get round to implementing feature 193 and 570 we can
adjust this to suit.

Fixes #571.