Realtime Plot shows last 60s, not last 10s
The realtime plot was originally set to only show the
last 10s of activity to reduce CPU overhead.

In reality, the CPU overhead for the plot is minimal,
especially when anti-aliasing is disabled.

This update modifies the plot to show the last 60s of
data, instead of the rather restrictive 10s.