Cleanup of Tacx Fortius support.
The Fortius code does not use a thread anymore, but simply uses singleshot
timers to perform the read from the device 70ms after the write. Also,
telemetry is pushed to the controller using Qt's signal-slot mechanism.

Running averages are computed using the new RunningAverage class, which
simply computes the average of the last n items added to the object.

Furthermore, I've tried to clean up the code to minimize the compile
dependencies of the files.