Remove reading of .notes from LTMPopup
Last vestiges of references to the old .notes file which
is now deprecated fully and replaced with a metadata tag
"Notes".

If a .notes file exists it will be read in and used to
initialize the Notes metadata tag -- this is pre-existing
functionality and not changed with this patch, which is
just related to LTM Popup.

There are no other know references to the old .notes file
left in the active code.

Fixes #179.