This patch adds an air density (rho) calculator feature to GoldenCheetah.
Users supply the estimator with measurements of temperature,
air pressure, and dew point, and the estimator uses Herman Wobus'
model for vapor pressure, plus the standard equations for
calculating air density, to estimate Rho. This feature is
useful for dialing in the Rho value in aerolab.

The patch hooks the estimator into the Tools menu as another dialog
box, similar to the CP estimator that's also under the tools menu.