Reinstate ride calendar
A rudimentary calendar, not customisable or resizable
and only every selects first ride for the day selected.
But it is the beginning of the widgets needed for the
new UI "look" discussed on the list.

Need to work on font sizes for Win/Mac and refining
the whole look and feel. Especially a sidebar widget
to put it in!

Fixes #369.