Overhaul gcconfig.pri.in and src.pro
Changes to src.pro are to make it so that anything that is set in src.pro
can be overwritten out of gcconfig.pri. This hopefully means that src.pro
will never have to be changed by someone wanting to compile the code using
a different setup to the main developers or on different platform.

Changes to gcconfig.pri.in are to allow the changes in src.pro to be used.
It also adds some new tweaks and documents how to use the changes in the file.