Fortius I/O is now done in it's own thread.
Fortius connection to usb is done on it's thread. Telemetry
is communicated to the controller using signals. The controller
calls the Fortius object indirectly using QMetaObject::invokeMethod,
which results in asynchronous (queued) slot calls. This means
that the methods called run on the thread on which the Fortius
object lives.