Link compress libs (-lz) by default
Although you should not link with local compress
libraries if you are using the QT supplied ones, it
has caused a lot of problems for newbies.

We now link with compress by default, we may see some
odd / incorrect behaviour with images and TrainingPeaks
upload if the libs mismatch but the developer can fix
that should it arise.

Fixes #668.
Fixes #6.