Refactoring CLucene
There are some issues with document deletion and support for
the latest release of CLucene which may be related to having
concurrent IndexWriters and IndexReaders.

This refactoring works fine on Linux. Pushed for safety for
any users that may be experiencing issues when deleting rides.

Further investigation is required on mac and windows builds.