Basic default for the home view
Having used the LTM and Home View for well
over a year I am proposing this default setup
for the home view.

It is just a PMC, Power tracker, CP chart and
a time in zone histogram.

No time and distance etc. But the user can add
these, the charts.xml is there to help.