Calendar sidebar cosmetics
It should be white to be in line with the rest of the
sidebars and have some semblance of coherence.

I also put a rectangle around the box so if the day
color is white you can still see there is a ride that day.

A summary ala the interval summary will come next summarising
either day, week or month for the ride selected.