UI Nits: RideNavigator margins to allow resize
The ride navigator has 0 contents margins which made
grabbing and resizing impossible in paned mode. This
was done to make it look beter when it is a ride list.

It now has a constructor argument to specify if its
being put into mainwindow as the ride list or its being
added to a view.