Fix ride navigator crash on search
The ride navigator crashed on search after the new
time column was added. This was caused by incorrectly
setting the groupBy member variable (-1 is used to
indicate no group by, it was set to -2).

In addition to fixing the code to set it to -1 the
array reference now does bounds checking to avoid
crashing for invalid values.