Mainain Events in LTM Sidebar
Add the ability to create and maintain events in
the LTMSidebar. This is a follow up to adding
support for Events in seasons.xml.

The last step is now to add support for annotating
charts to show the events for the date range selected
and possibly for the seasons within that date range.