TreeMap Chart - refactor away LTM
The treemap chart was developed as a variation of an LTM
chart and as a results has the same chart settings etc.

It has now been refactored to be a standalone chart with
no need to re-use the LTM settings and code.

This is also in preparation to support click through on
the cells to show a popup of rides.