Tidy up Preferences Dialog
Big refactor of the code since it was really messy
and also moved to a Mac style config dialog with
a bit of Mac UI polish.

Moved many of the pages around to make it more user
friendly.

Removed the fortius and proxy code altogether
and removed the measures config since we will need to
look at this in v3.1 and don't want to have to add
backward compatibility support for users that change it.

Fixes #358.