Fix Split Ride SEGV
Appears to be related to add/removing labels during a paint event,
which is pretty bad practice anyway. To get past this crash the
labels have been removed -- the chart still shades areas to show
the parts of the ride that will be split out, but no longer shows
the number for each segment.

Fixes #379.