Chart Dates: Part 2 of 3 update
.. added a 'This' option to choose a date range
such as this month, week, year. I also added the ability
to specift a 'prior' period.

So in July you could plot for this month prior 1 to choose
a date range for June.

This allows folks to have a static setup of charts for this
this-1 and this-2 months and data to change over time.