UI Nits: Reveal controls on black bg
.. black controls on a black background might
be cool if you play in a plutonium rock band
from the Gagrakacka Mind Zones, but for the
rest of us its helpful if the controls are
slightly more visible.

A bit of hacking with style sheets and the controls
now change color to contrast with whatever background
color is selected.