Wahoo Fitness API Abstraction Part 1
Just the framework for the abstraction layer. I
wanted to commit this since its the blank state
and clean.

Now to implement;
1 hasBTLESupport and friends
2 scanForDevices
3 connect/disconnect device
4 getRealtimeData
5 setLoad/setGrade

Once thats done should be easy to write a controller
and runner for TrainView.

Last step will be to add a configuration pane -- but
might wait until I have a BTLE HR strap to include
the 'multi-device' stuff.