Code Cleanup: Valgrind Memory
Not sure why, but setting the text on a treewidget item seems to
cost a lot of memory. So we don't bother since it is never used.

In general the valgrind output is ok, quite surprised. There are
very few non-widget items that are alloced but never free'd in the
GC code.